REF: share DTA/TDA/PA arithmetic methods
