+++ pandas/core/arrays/datetimelike.py
@@ -72,6 +72,7 @@ from pandas.util._decorators import (
 from pandas.util._exceptions import find_stack_level
 
 from pandas.core.dtypes.common import (
+    DT64NS_DTYPE,
     is_all_strings,
     is_categorical_dtype,
     is_datetime64_any_dtype,
@@ -89,7 +90,10 @@ from pandas.core.dtypes.common import (
     is_unsigned_integer_dtype,
     pandas_dtype,
 )
-from pandas.core.dtypes.dtypes import ExtensionDtype
+from pandas.core.dtypes.dtypes import (
+    DatetimeTZDtype,
+    ExtensionDtype,
+)
 from pandas.core.dtypes.missing import (
     is_valid_na_for_dtype,
     isna,
@@ -113,6 +117,7 @@ from pandas.core.arrays._mixins import (
 import pandas.core.common as com
 from pandas.core.construction import (
     array as pd_array,
+    ensure_wrapped_if_datetimelike,
     extract_array,
 )
 from pandas.core.indexers import (
@@ -1082,26 +1087,123 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
     __divmod__ = make_invalid_op("__divmod__")
     __rdivmod__ = make_invalid_op("__rdivmod__")
 
+    @final
     def _add_datetimelike_scalar(self, other):
         # Overridden by TimedeltaArray
-        raise TypeError(f"cannot add {type(self).__name__} and {type(other).__name__}")
+        if not is_timedelta64_dtype(self.dtype):
+            raise TypeError(
+                f"cannot add {type(self).__name__} and {type(other).__name__}"
+            )
 
-    _add_datetime_arraylike = _add_datetimelike_scalar
+        from pandas.core.arrays import DatetimeArray
 
-    def _sub_datetimelike_scalar(self, other):
-        # Overridden by DatetimeArray
         assert other is not NaT
-        raise TypeError(f"cannot subtract a datelike from a {type(self).__name__}")
+        other = Timestamp(other)
+        if other is NaT:
+            # In this case we specifically interpret NaT as a datetime, not
+            # the timedelta interpretation we would get by returning self + NaT
+            result = self.asi8.view("m8[ms]") + NaT.to_datetime64()
+            return DatetimeArray(result)
 
-    _sub_datetime_arraylike = _sub_datetimelike_scalar
+        i8 = self.asi8
+        result = checked_add_with_arr(i8, other.value, arr_mask=self._isnan)
+        result = self._maybe_mask_results(result)
+        dtype = DatetimeTZDtype(tz=other.tz) if other.tz else DT64NS_DTYPE
+        return DatetimeArray(result, dtype=dtype, freq=self.freq)
 
+    @final
+    def _add_datetime_arraylike(self, other):
+        if not is_timedelta64_dtype(self.dtype):
+            raise TypeError(
+                f"cannot add {type(self).__name__} and {type(other).__name__}"
+            )
+
+        # At this point we have already checked that other.dtype is datetime64
+        other = ensure_wrapped_if_datetimelike(other)
+        # defer to DatetimeArray.__add__
+        return other + self
+
+    @final
+    def _sub_datetimelike_scalar(self, other: datetime | np.datetime64):
+        if self.dtype.kind != "M":
+            raise TypeError(f"cannot subtract a datelike from a {type(self).__name__}")
+
+        self = cast("DatetimeArray", self)
+        # subtract a datetime from myself, yielding a ndarray[timedelta64[ns]]
+
+        # error: Non-overlapping identity check (left operand type: "Union[datetime,
+        # datetime64]", right operand type: "NaTType")  [comparison-overlap]
+        assert other is not NaT  # type: ignore[comparison-overlap]
+        other = Timestamp(other)
+        # error: Non-overlapping identity check (left operand type: "Timestamp",
+        # right operand type: "NaTType")
+        if other is NaT:  # type: ignore[comparison-overlap]
+            return self - NaT
+
+        try:
+            self._assert_tzawareness_compat(other)
+        except TypeError as err:
+            new_message = str(err).replace("compare", "subtract")
+            raise type(err)(new_message) from err
+
+        i8 = self.asi8
+        result = checked_add_with_arr(i8, -other.value, arr_mask=self._isnan)
+        result = self._maybe_mask_results(result)
+        return result.view("timedelta64[ns]")
+
+    @final
+    def _sub_datetime_arraylike(self, other):
+        if self.dtype.kind != "M":
+            raise TypeError(f"cannot subtract a datelike from a {type(self).__name__}")
+
+        if len(self) != len(other):
+            raise ValueError("cannot add indices of unequal length")
+
+        self = cast("DatetimeArray", self)
+        other = ensure_wrapped_if_datetimelike(other)
+
+        try:
+            self._assert_tzawareness_compat(other)
+        except TypeError as err:
+            new_message = str(err).replace("compare", "subtract")
+            raise type(err)(new_message) from err
+
+        self_i8 = self.asi8
+        other_i8 = other.asi8
+        arr_mask = self._isnan | other._isnan
+        new_values = checked_add_with_arr(self_i8, -other_i8, arr_mask=arr_mask)
+        if self._hasna or other._hasna:
+            np.putmask(new_values, arr_mask, iNaT)
+        return new_values.view("timedelta64[ns]")
+
+    @final
     def _sub_period(self, other: Period):
-        # Overridden by PeriodArray
-        raise TypeError(f"cannot subtract Period from a {type(self).__name__}")
+        if not is_period_dtype(self.dtype):
+            raise TypeError(f"cannot subtract Period from a {type(self).__name__}")
 
+        # If the operation is well-defined, we return an object-dtype ndarray
+        # of DateOffsets.  Null entries are filled with pd.NaT
+        self._check_compatible_with(other)
+        asi8 = self.asi8
+        new_data = asi8 - other.ordinal
+        new_data = np.array([self.freq.base * x for x in new_data])
+
+        if self._hasna:
+            new_data[self._isnan] = NaT
+
+        return new_data
+
+    @final
     def _add_period(self, other: Period):
-        # Overridden by TimedeltaArray
-        raise TypeError(f"cannot add Period to a {type(self).__name__}")
+        if not is_timedelta64_dtype(self.dtype):
+            raise TypeError(f"cannot add Period to a {type(self).__name__}")
+
+        # We will wrap in a PeriodArray and defer to the reversed operation
+        from pandas.core.arrays.period import PeriodArray
+
+        i8vals = np.broadcast_to(other.ordinal, self.shape)
+        parr = PeriodArray(i8vals, freq=other.freq)
+        return parr + self
 
     def _add_offset(self, offset):
         raise AbstractMethodError(self)
@@ -1116,9 +1218,9 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         """
         if isna(other):
             # i.e np.timedelta64("NaT"), not recognized by delta_to_nanoseconds
-            new_values = np.empty(self.shape, dtype="i8")
+            new_values = np.empty(self.shape, dtype="i8").view(self._ndarray.dtype)
             new_values.fill(iNaT)
-            return type(self)(new_values, dtype=self.dtype)
+            return type(self)._simple_new(new_values, dtype=self.dtype)
 
         # PeriodArray overrides, so we only get here with DTA/TDA
         # error: "DatetimeLikeArrayMixin" has no attribute "_reso"
@@ -1139,7 +1241,9 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
             new_values, dtype=self.dtype, freq=new_freq
         )
 
-    def _add_timedelta_arraylike(self, other):
+    def _add_timedelta_arraylike(
+        self, other: TimedeltaArray | npt.NDArray[np.timedelta64]
+    ):
         """
         Add a delta of a TimedeltaIndex
 
@@ -1152,11 +1256,8 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         if len(self) != len(other):
             raise ValueError("cannot add indices of unequal length")
 
-        if isinstance(other, np.ndarray):
-            # ndarray[timedelta64]; wrap in TimedeltaIndex for op
-            from pandas.core.arrays import TimedeltaArray
-
-            other = TimedeltaArray._from_sequence(other)
+        other = ensure_wrapped_if_datetimelike(other)
+        other = cast("TimedeltaArray", other)
 
         self_i8 = self.asi8
         other_i8 = other.asi8
@@ -1200,12 +1301,27 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         result.fill(iNaT)
         return result.view("timedelta64[ns]")
 
-    def _sub_period_array(self, other):
-        # Overridden by PeriodArray
-        raise TypeError(
-            f"cannot subtract {other.dtype}-dtype from {type(self).__name__}"
+    @final
+    def _sub_period_array(self, other: PeriodArray) -> npt.NDArray[np.object_]:
+        if not is_period_dtype(self.dtype):
+            raise TypeError(
+                f"cannot subtract {other.dtype}-dtype from {type(self).__name__}"
+            )
+
+        self = cast("PeriodArray", self)
+        self._require_matching_freq(other)
+
+        new_values = checked_add_with_arr(
+            self.asi8, -other.asi8, arr_mask=self._isnan, b_mask=other._isnan
         )
 
+        new_values = np.array([self.freq.base * x for x in new_values])
+        if self._hasna or other._hasna:
+            mask = self._isnan | other._isnan
+            new_values[mask] = NaT
+        return new_values
+
+    @final
     def _addsub_object_array(self, other: np.ndarray, op):
         """
         Add or subtract array-like of DateOffset objects
@@ -1301,6 +1417,8 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
             # as is_integer returns True for these
             if not is_period_dtype(self.dtype):
                 raise integer_op_not_supported(self)
+            # FIXME: inconsistent with the scalar behavior which multiplies
+            #  by per.freq.n
             result = self._time_shift(other)
 
         # array-like others
@@ -1316,6 +1434,8 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         elif is_integer_dtype(other_dtype):
             if not is_period_dtype(self.dtype):
                 raise integer_op_not_supported(self)
+            # FIXME: inconsistent with the scalar behavior which multiplies
+            #  by per.freq.n
             result = cast("PeriodArray", self)._addsub_int_array_or_scalar(
                 other, operator.add
             )
