+++ pandas/core/arrays/period.py
@@ -85,7 +85,10 @@ if TYPE_CHECKING:
         NumpyValueArrayLike,
     )
 
-    from pandas.core.arrays import DatetimeArray
+    from pandas.core.arrays import (
+        DatetimeArray,
+        TimedeltaArray,
+    )
 
 
 _shared_doc_kwargs = {
@@ -548,6 +551,7 @@ class PeriodArray(dtl.DatelikeOps, libperiod.PeriodMixin):
                 "`freq` argument is not supported for "
                 f"{type(self).__name__}._time_shift"
             )
+        # TODO: similar to addsub_int_array_or_scalar?
         values = self.asi8 + periods * self.freq.n
         if self._hasna:
             values[self._isnan] = iNaT
@@ -712,46 +716,6 @@ class PeriodArray(dtl.DatelikeOps, libperiod.PeriodMixin):
     # ------------------------------------------------------------------
     # Arithmetic Methods
 
-    def _sub_period(self, other: Period):
-        # If the operation is well-defined, we return an object-Index
-        # of DateOffsets.  Null entries are filled with pd.NaT
-        self._check_compatible_with(other)
-        asi8 = self.asi8
-        new_data = asi8 - other.ordinal
-        new_data = np.array([self.freq.base * x for x in new_data])
-
-        if self._hasna:
-            new_data[self._isnan] = NaT
-
-        return new_data
-
-    def _sub_period_array(self, other):
-        """
-        Subtract a Period Array/Index from self.  This is only valid if self
-        is itself a Period Array/Index, raises otherwise.  Both objects must
-        have the same frequency.
-
-        Parameters
-        ----------
-        other : PeriodIndex or PeriodArray
-
-        Returns
-        -------
-        result : np.ndarray[object]
-            Array of DateOffset objects; nulls represented by NaT.
-        """
-        self._require_matching_freq(other)
-
-        new_values = algos.checked_add_with_arr(
-            self.asi8, -other.asi8, arr_mask=self._isnan, b_mask=other._isnan
-        )
-
-        new_values = np.array([self.freq.base * x for x in new_values])
-        if self._hasna or other._hasna:
-            mask = self._isnan | other._isnan
-            new_values[mask] = NaT
-        return new_values
-
     def _addsub_int_array_or_scalar(
         self, other: np.ndarray | int, op: Callable[[Any, Any], Any]
     ) -> PeriodArray:
@@ -810,7 +774,9 @@ class PeriodArray(dtl.DatelikeOps, libperiod.PeriodMixin):
 
         return super()._add_timedeltalike_scalar(other)
 
-    def _add_timedelta_arraylike(self, other):
+    def _add_timedelta_arraylike(
+        self, other: TimedeltaArray | npt.NDArray[np.timedelta64]
+    ):
         """
         Parameters
         ----------
